fix: don't use wait-for-it to start the crawler
For some reason, using args results in an error:
"args does not match original args defined by env-injector
application=env-injector component=akv2k8s".

Maybe it's because of the environment variable that holds
different values at different stages of the deploy?

Anyway, we don't strictly need wait-for-it because Elasticsearch
is supposed to be up and we are ok to fail if it happens to be down.